pyproject.toml
[tool.poetry]
name = "hackme"
version = "0.1.0"
description = "A hackathon registration tool"
authors = ["p1utoze <1ds21ai001@dsce.edu.in>"]
license = "Apache-2.0"

[tool.poetry.scripts]
start-app = "app.main:app"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
uvicorn = "0.22.0"
gunicorn = "20.1.0"
Pillow = "10.1.0"
fastapi = {extras = ["standard"], version = "0.95.2"}
fastapi-sso = "^0.9.0"
firebase = "^4.0.1"
firebase-admin = "^6.2.0"
pre-commit = "^3.5.0"
qrcode = "7.4"
pandas = "2.0.2"
python-multipart = "0.0.5"
Jinja2 = "^3.1.2"
poetry-dotenv = "^0.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
